(CFACT-227) Move logging and boost-nowide into leatherman
This vendors boost-nowide instead of making it a submodule, in order
to avoid recursive submodules.

I also pulled basic CFLAG setting out of cfacter, since leatherman
needs the logic to build itself as C++11. This means cfacter now
relies on leatherman to set up its core cflags as well (to ensure we
don't screw up our ABI between leatherman libraries and the apps that
use it)