(CFACT-232) Get unit tests to actually run correctly
Previously, we were always generating an executable with no
tests. Catch uses the "Create a global object and let its constructor
register it" pattern for tracking tests. This falls over when tests
are in static libraries, since the linker discards the object files as
"unused" without realizing that constructors need to be called.

This commit fixes up the linker flags to ensure the tests are actually
brought in, and does some fidgeting with library order to make
everything sane/stable.