(CFACT-235) Clean up handling of defoptions for use by containing projects
The defoptions that were intended to be shared with containing
projects had been tossed at the top of `cflags.cmake`. It turns out we
really want to think about these as separate things, so they've been
split out.

Also now we only pull in cflags and options if we're top-level. When
leatherman is embedded we want to rely on the calling project to deal
with this stuff.