(LTH-5) Add Boost.Chrono explicit dependency
Boost.Log has an implicit dependency on Boost.Chrono. When we statically link,
this gets resolved at compile time and isn't an issue. On most platforms when
we dynamically link, the Boost library location is in LD_LIBRARY_PATH, and
again Boost.Chrono is found.

On Solaris neither of those things are true, which exposed a runtime
dependency on libboost_chrono.so. Resolve this by explicitly depending on the
Chrono library in Leatherman's Logging library.