(LTH-22) Add terminal colors on Windows
Uses SetConsoleTextAttribute to add color to log messages on Windows when
writing to a stdout or stderr stream.

The colorize functions were changed to account for Windows updating a
global rather than using a control string, so they now take an ostream to
colorize. Tests were updated to account for that.

To add color on Windows on a portion of a log line, log formatting had to
be moved down to the log backend. Adding line numbers were moved to an
attribute that is written as part of the backend, so it can no longer be
tested by creating a custom backend; therefore the line number tests were
removed.