(LTH-34) Cope with rapidjson Size() assertion
rapidjson Size() method asserts that the caller object is an array

This causes the unit tests to fail on Ubuntu in due to the above
assertion. Note that they all execute and pass on OS X.

With this commit we only call the Size() method for array entries; for
object entries we use MemberCount(), whereas for scalars we simply
return 0.