(LTH-41) Add function to specify curl protocols
This commit adds a wrapper to specify CURLOPT_PROTOCOLS per a curl
request. If no protocol is defined, it will default to all protocols. If
a protocol is defined but one is used that wasn't specified (like using
http when https was the only protocol enabled), it will fail with an
http exception saying that an unsupported protocol was used.