(LTH-41) Add curl protocol settings tests
This commit adds a few basic tests around setting what protocols will be
used for a given curl request. It checks to make sure that if you
specify a protocol, it is able to make that request. It also has a
failure case to ensure that if you request a specific protocol but try
to make a different one, it will throw an exception.