(maint) Refactor how _client_protocols default value is set
Prior to this commit, _client_protocols was uninitialized when declared
and would be given a default in the method that uses that variable if it
was still unset. This commit updates that to actually give the default
value to _client_protocols when it is first initialized. If the user
wants that var to be something other than the default, it will be over
written when the helper method is called.