(FACT-891) Always assume dynamic_library::load is first load
It's not possible on all platforms to check if a library has been
loaded. Instead, the right thing to do is use
dynamic_library::find_by_sybmol and only call load if that
failed. Since that's the new recommended API usage, and since Facter
already does the right thing, load can now unconditionally flag the
library as first_load.