(maint) Port fix in facter#stable to Leatherman execution library.
This ports the fix for FACT-1158 (failure to properly expand commands for Ruby
command execution) that is currently in facter's stable branch to Leatherman.

The fix is to properly quote expanded commands that contain spaces.
Additionally, a new option is added to preserve the arguments so that they
aren't automatically quoted on Windows.