(LTH-43) Add getWithDefault to JsonContainer
Before this patch, if users of JsonContainer wanted to try to find
an entry in a JSON object and use a default value if the entry was not
found, they would have to perform a two-step process of checking for the
key and then either getting the key or using the default value. This is
more work than is strictly necessary and leads to superfluous repetition
of this logic. With this patch applied, that logic is centralized in the
`getWithDefault` method, and users can call that function with their
default value instead of implementing the logic themselves.