(LTH-48) Return exit_code from execute command
The execute commands previously had no way to return the exit code from
the child process without throwing an exception. Many use cases include
examining the exit code of a failed child process as normal behavior, so
we should provide a way to get it without using exceptions. Add it to
the return tuple to make it available to callers.