(LTH-49) Add child process stdin support to execute
Adds support for passing a string to the `execute` commands with the full
stream to pass to stdin. Input will be interleaved with output as
controlled by the invoked process requesting new input.

Also adds a basic `cat` utility - as one can't be expected on Windows -
with additional options for more complex program behavior.