(LTH-30) Use threadid to avoid named pipes colliding
Named pipe names can collide in the same process if Leatherman.Execution
is statically linked. Separate DLLs that each statically link and use it,
that invoke a process concurrently in separate threads could potentially
use the same name. Use thread ID instead, as that should be unique for a
single process invocation.