(maint) Fix occasionally skipping open pipes
Previously a bug existed where we could drop trailing data on stderr when
overlapping i/o occurs in Leatherman.Execution's execute command. After
stdout signals that it's closed, it won't attempt to read further from
stderr. If there's pending i/o on stderr, that i/o is lost.

I haven't seen this manifest in practice without changing the code, but it
has made the code fragile to work with.