(LTH-51) Protect against potential named pipe re-use
There's a scenario using timeouts where we could release the invoking
end of a named pipe but the other end doesn't release using
Leatherman.Execution. Then the invoking thread shuts down and another
with the same thread id is started and reconnects to the existing named
pipe.

Use the process id and a random number to make that scenario highly
unlikely. This follows a similar pattern used in Chromium:
https://chromium.googlesource.com/chromium/src.git/+/master/base/sync_socket_win.cc#46