(LTH-53) Don't set explicit Content-Length
Curl expects to receive the length of uploaded content as an
option (either CURLOPT_POSTFIELDSIZE or CURLOPT_INFILESIZE for a post or
put respectively). It will then set Content-Length when it's
appropriate. Setting the explicit header causes the header to get sent
even when inappropriate (such as when a redirect causes a post to get
resent as a get).