(LTH-52) Fix memory use of getRaw
Many places use getRaw to get access to the raw rapidjson
Document type, but then treat it as a reference and drop ownership.
This doesn't crash and burn only because the MemoryPoolAllocator never
releases memory. getRaw also creates unnecessary copies.

Change to return a const ref to the original object.