(maint) Fix logging_sinks test for color
The logging_sinks test for colorization was original written when
logging was formatted via a a custom formatter. The test created a
custom sink to consume the formatted text, and verified that it
contained the color strings queried from the log level.

In a rewrite to make colored output work on Windows, logging was changed
to use a custom sink so it can configure the console color when writing
to the output stream. The tests were rewritten, but due to test ordering
all the platforms Leatherman unit tests were run on passed because
logging reported color types with colorization disabled. The tests
should have failed because we replaced the sink that did colorization
with one capturing the output for testing.

Testing with streams is already covered by existing tests, so remove the
test using a custom sink. Leatherman.Logging formatting does not apply
to other custom sinks.

Keep tests of the `operator>>` for `log_level` and `on_message`.