(LTH-81) Drop dependencies when using shared libraries
When Leatherman is built as a set of shared libraries, all required
dependencies are accounted for in the binary files. However the
LEATHERMAN_LIBS and LEATHERMAN_*_LIBS variables still included dependent
libraries, which could be static libraries. When the dependencies are
static libraries, linking using LEATHERMAN_*_LIBS can result in missing
symbols.

Update Leatherman to leave dependencies out of LEATHERMAN_*_LIBS when
built with shared libraries.