Update example for how to use i18n features
This commit adds a bit more context to the readme for using the i18n
features of Leatherman. It adds the work needed to get it compiling in a
projects CMakeLists.txt file, and also gives a small example on the few
lines of code needed for translating and formatting strings.