(maint) Fix 32-bit compilation
A change to using 64-bit integers in the Ruby API assumed size_t would
always be 64-bit. That breaks 32-bit compilation. Restore explicitly
casting to size_t.

Getting a size and expecting it to match size_t is also a common
pattern. Add a helper that does bounds-checking for that cast.