(LTH-96) Ensure LEATHERMAN_LOCALES rules are generated
Previously two CMake runs would be required to add new locales via
LEATHERMAN_LOCALES. The first would generate the rules to build the .po
files, but not rules to compile those to .mo files. The second would add
the rules to make .mo files. Ensure the rules to make .mo files are
added on the first run.