(LTH-96) Prevent locale files from being cleaned
Previously `make clean` would delete generated locale files. Prevent
that from happening, as the file generation is only meant as a
bootstrapping step; .po files are then checked in and updated over time.

This fix only works for `make`, as CMake's documentation notes it can't
be implemented for other generators.