(LTH-107) Add inherit_locale option to execute
The `execute` function was originally implemented for Facter calls to
command-line tools, where it wants consistently parseable output. For
that purpose, it always overrode LC_ALL and LANG.

We now use `execute` for other purposes, so allow inheriting the locale
from the executing process.

Also ensure Windows execution supports Unicode in the environment.