(PCP-556) Logging backend doesn't consume records with lvl==none
The Logging logger (logging::helper) is currently in charge of checking
the severity (compared to the configured log level) before creating a new
boost.log record. The sink backend (logging::color_writer) does
not check that when it consumes a record (color_writer::consume). Such
behaviour does not cope well with records created by severity loggers
other than the Logging logger, as the severity value gets ignored.

With this commit we deal with the case where a record created with a
third party severity logger has severity equals to 'none'; in such
case, the record will be discarded.