(LTH-116) Handle null characters in JSON strings
Before this patch leatherman could not handle null characters (`\u0000`)
in JSON keys/values because it used the single argument `std::string`
constructor to create `std::string` representations of those keys/values.
The problem was that the null character is treated as the end of the
string by that constructor, so any key/value containing the null character
was truncated at that character when represented as the `std::string`.
The fix is to use the two argument constructor of `std::string` where the
second argument specifies the length of the value to store in the string.
The respective length is readily obtainable from the JSON parsing library.