(LTH-120) Rename create_new_process_group -> create_detached_process
This patch renames the `create_new_process_group` execution option to
`create_detached_process`.
The intention of the option was to create a process in such a way it is
detached from its parent in terms of their lifespan. On Windows, which
was the only platform where the option was implemented, the child process
would be killed when the parent process exited in case the option was
not specified. On the contrary the child can outlive its parent when the
option is specified.
The renaming of the option is a preparation for generalization of the
concept to other OSs where it is also applicable (notably Solaris) but
where the original name would not make sense or would be misleading.