(LTH-125) Ensure file operations around exec are thread-safe
Use a mutex to ensure that modifying file descriptors does not occur
concurrently if multiple calls to execute are done at once.

Without locking, one thread could mark the descriptors close-on-exit
just prior to vfork and another thread just complete a vfork and mark
the same descriptors as not close-on-exec just in time for the first
thread to exec without them flagged.