(LTH-126) Make it possible to use fork() on demand
Traditionally leatherman::execution used `vfork()` for the child process
creation because of its higher performance and lower virtual memory
requirements. But this turned out to be problematic (deadlock prone) in
multi-threaded processes on at least Solaris.
So in this patch a new execution option - `thread_safe` - is introduced,
which when specified causes leatherman::execution to use slower but safer
`fork()` for the child process creation. If the option is not specified
the more performant but more dangerous `vfork()` is used to preserve backward
compatibility.