(LTH-126) Partially revert commits related to LTH-125
This reverts the pieces of the LTH-125 related commits which replaced closing
file descriptors in the child process by setting & unsetting `FD_CLOEXEC` flag
on those file descriptors in the parent process.
In addition to that the function which returns the highest possible file
descriptor number - `get_max_descriptor_limit()` - is called in the parent
process and its result is passed as a parameter down to the function which
actually closes the file descriptors in the child process. This is to avoid
calling the function in the vfork context which may be deadlock prone.
With the `get_max_descriptor_limit()` function call out of the way, we are
calling only one extra function (compared to the reverted code) in the vfork
context - namely: `close()`.
Calling the `close()` should not be any more deadlock prone than calling the
other functions we do call in the vfork context, e.g. `dup2()`.