(LTH-128) Use less Boost.Log magic
On AIX, when used with pxp-agent, logging doesn't work because the
log-level (severity) isn't correctly stored by the severity logger. This
is strange as it does work with Facter; I suspect the cause is due to
differences in linking with a .exe (pxp-agent) vs a .so (libfacter)
against libboost_logging.so. Work around the issue in Boost.Log by using
less of its magic, and instead using an explicit attribute for severity.