(LTH-135) Make gettext_compile respect LOCALE env vars
This commit fixes an issue where if the user set LEATHERMAN_LOCALE_VAR
and LEATHERMAN_LOCALE_INSTALL to determine where leatherman should look
for translation files, the install of those files from consuming
projects would not respect those vars and install them in the default
location, which may not correspond to the dir specified with those vars.
Now if the vars are set, leatherman will both install to the specificed
directory and read translations from it at runtime.