(LTH-149) Handle errors if child has exited before all writing finished
In pxp-agent testing, we periodically see `failed to run: stdin pipe i/o
failed: Broken pipe (32)` on AIX. This occurs when the executed script
doesn't consume stdin, and exits before we write stdin to the pipe.
Catch EPIPE on stdin and consider the pipe closed.

This new behavior is opt-in via the allow_stdin_unread flag.