(FACT-1919) Add environment::get_int
Add a new helper function to simplify getting integer environment
variables from the system.

The function parameters are the environment variable name, and the
default value to return. If the environment variable to get does not
exist, or cannot be cleanly coverted to an integer, the default value is
returned instead.