(LTH-168) Get DWORD values from Windows Registry
This commit extends the private `get_regvalue` function to also
resolve DWORD values from the registry.

For external usage, the `get_registry_dword` function is added, which
behaves similar to `get_registry_string` but works on DWORDs and
returns an unsigned long value.

Getting the value as DWORD, converting to wstring, then converting
again to unsigned long is redundant but I didn't want to write another
function for this specific functionality.