(PE-36763) Add backwards-compat support for CURLOPT_PROTOCOLS_STR
Update the leatherman curl client to support usage of
CURLOPT_PROTOCOLS_STR when it is available. Versions of cURL earlier
than 8 will not have CURLOPT_PROTOCOLS_STR, so these changes allow usage
of CURLOPT_PROTOCOLS if the new string version of the function is not
available.

The _client_protocols param of the client context is now a string, and
there is a second implementation of set_supported_protocols that takes a
string in the form CURLOPT_PROTOCOLS_STR expects.

Regardless of which version of cURL leatherman is compiled with, both
the bitmask and string versions of set_supported_protocols perform the
same logic. However the set_client_protocols function - the function
that actually calls CURLOPT_PROTOCOLS or CURLOPT_PROTOCOLS_STR -
differs: with older cURL it translates the strings back to bitmasks,
with cURL newer than version 8 it sends the string directly to
CURLOPT_PROTOCOLS_STR.

Structurally the code was written in an attempt to 1. move everything
towards the use of strings with CURLOPT_PROTOCOLS_STR and 2. to avoid
differently compiled code between versions as much as possible.