(maint) Remove -DLEATHERMAN_SHARED=OFF test
Everything we use leatherman to build with uses -DLEATHERMAN_SHARED=true
(linking to it as a shared library), so we don't need to test that
cmake flag.

New changes ended up breaking for the mock_curl used for testing with
that flag on, and we simply don't have the priority to fix it. So just
remove that test